LPD-34375 Move Undo/Redo related eventListeners to WebContentVairant
