+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/core/components/PageRenderer/WebContentVariant.es.js
@@ -5,7 +5,7 @@
 
 import ClayLayout from '@clayui/layout';
 import classnames from 'classnames';
-import React, {forwardRef, useRef} from 'react';
+import React, {forwardRef, useEffect, useRef} from 'react';
 import {useDrop} from 'react-dnd';
 
 import {EVENT_TYPES} from '../../../custom/form/eventTypes';
@@ -34,7 +34,7 @@ export const Column = forwardRef(
 		},
 		ref
 	) => {
-		const {portletId} = useFormState();
+		const {defaultLanguageId, portletId} = useFormState();
 
 		const addr = {
 			'data-ddm-field-column': index,
@@ -45,6 +45,43 @@ export const Column = forwardRef(
 		const firstField = column.fields[0];
 		const isFieldSetOrGroup = firstField?.type === 'fieldset';
 		const isFieldSet = firstField?.ddmStructureId && isFieldSetOrGroup;
+
+		const defaultLanguageIdRef = useRef(defaultLanguageId);
+		const dispatch = useForm();
+
+		useEffect(() => {
+			if (defaultLanguageIdRef.current !== defaultLanguageId) {
+				defaultLanguageIdRef.current = defaultLanguageId;
+				dispatch({type: EVENT_TYPES.HISTORY.RESET});
+			}
+		}, [defaultLanguageId, dispatch]);
+
+		useEffect(() => {
+			const goToHandler = ({step}) => {
+				dispatch({step, type: EVENT_TYPES.HISTORY.GOTO});
+			};
+			const handleStoreState = () => {
+				dispatch({type: EVENT_TYPES.HISTORY.ADD});
+			};
+			const undoHandler = () => {
+				dispatch({type: EVENT_TYPES.HISTORY.PREV});
+			};
+			const redoHandler = () => {
+				dispatch({type: EVENT_TYPES.HISTORY.NEXT});
+			};
+
+			Liferay.on('journal:goto', goToHandler);
+			Liferay.on('journal:undo', undoHandler);
+			Liferay.on('journal:redo', redoHandler);
+			Liferay.on('journal:storeState', handleStoreState);
+
+			return () => {
+				Liferay.detach('journal:goto', goToHandler);
+				Liferay.detach('journal:undo', undoHandler);
+				Liferay.detach('journal:redo', redoHandler);
+				Liferay.detach('journal:storeState', handleStoreState);
+			};
+		}, [dispatch]);
 
 		return (
 			<ClayLayout.Col
